Implemented parallel firing of neurons, multi threaded network batches, guava and better logging (log4j2)

+++ src/main/java/de/jklotz/neuralnetwork/activation/ActivationFunction.java
@@ -16,9 +16,9 @@ public interface ActivationFunction extends Serializable {
     Softmax SOFTMAX = new Softmax();
 
 
-    double activate(Neuron neuron);
+    float activate(Neuron neuron);
 
-    double derivative(Neuron neuron);
+    float derivative(Neuron neuron);
 
     default boolean isStochasticDerivative() {
         return false;
