+++ src/main/java/de/jklotz/neuralnetwork/activation/Identity.java
@@ -5,12 +5,12 @@ import de.jklotz.neuralnetwork.network.Neuron;
 public class Identity implements ActivationFunction {
 
     @Override
-    public double activate(Neuron neuron) {
+    public float activate(Neuron neuron) {
         return neuron.getPreActivationValue();
     }
 
     @Override
-    public double derivative(Neuron neuron) {
+    public float derivative(Neuron neuron) {
         return 1;
     }
