+++ src/main/java/de/jklotz/neuralnetwork/activation/rectifier/ExponentialRectifier.java
@@ -5,20 +5,20 @@ import de.jklotz.neuralnetwork.network.Neuron;
 public class ExponentialRectifier extends ParametricRectifier {
 
     public ExponentialRectifier() {
-        this(0.01);
+        this(0.01f);
     }
 
-    public ExponentialRectifier(double leakiness) {
+    public ExponentialRectifier(float leakiness) {
         this.leakiness = leakiness;
     }
 
     @Override
-    public double activate(Neuron neuron) {
-        return neuron.getPreActivationValue() < 0 ? (leakiness * (Math.exp(neuron.getPreActivationValue()) - 1.0)) : neuron.getPreActivationValue();
+    public float activate(Neuron neuron) {
+        return neuron.getPreActivationValue() < 0 ? (float) (leakiness * (Math.exp(neuron.getPreActivationValue()) - 1.0f)) : neuron.getPreActivationValue();
     }
 
     @Override
-    public double derivative(Neuron neuron) {
-        return neuron.getPreActivationValue() < 0 ? (neuron.getValue() + leakiness) : 1.0;
+    public float derivative(Neuron neuron) {
+        return neuron.getPreActivationValue() < 0 ? (neuron.getValue() + leakiness) : 1.0f;
     }
 }
