+++ src/main/java/de/jklotz/neuralnetwork/activation/rectifier/SwishRectifier.java
@@ -8,13 +8,13 @@ public class SwishRectifier extends Rectifier {
 
 
     @Override
-    public double activate(Neuron neuron) {
+    public float activate(Neuron neuron) {
         return neuron.getPreActivationValue() * ActivationFunction.SIGMOID.activate(neuron);
     }
 
     @Override
-    public double derivative(Neuron neuron) {
-        double z = SIGMOID.activate(neuron);
-        return z * (1.0 + neuron.getPreActivationValue() * (1.0 - z));
+    public float derivative(Neuron neuron) {
+        float z = SIGMOID.activate(neuron);
+        return z * (1.0f + neuron.getPreActivationValue() * (1.0f - z));
     }
 }
