+++ src/main/java/de/jklotz/neuralnetwork/learning/LearningStrategy.java
@@ -6,6 +6,6 @@ public interface LearningStrategy {
 
     void learn();
 
-    double networkError();
+    float networkError();
 
 }
