+++ src/main/java/de/jklotz/neuralnetwork/network/Synapse.java
@@ -8,17 +8,17 @@ public class Synapse implements Serializable {
 
     public Neuron target;
 
-    public double weight;
+    public float weight;
 
-    public double change;
+    public float change;
 
-    public Synapse(Neuron source, Neuron target, double weight) {
+    public Synapse(Neuron source, Neuron target, float weight) {
         this.source = source;
         this.target = target;
         this.weight = weight;
     }
 
     public Synapse(Neuron source, Neuron target) {
-        this(source, target, Math.random());
+        this(source, target, (float) Math.random());
     }
 }
