+++ src/main/java/de/jklotz/neuralnetwork/normalization/Gaussian.java
@@ -3,20 +3,20 @@ package de.jklotz.neuralnetwork.normalization;
 public class Gaussian implements Normalization {
 
     @Override
-    public double[] normalize(double[] data) {
-        double[] result = new double[data.length];
-        double sum = 0.0;
-        for (double datum : data) {
+    public float[] normalize(float[] data) {
+        float[] result = new float[data.length];
+        float sum = 0;
+        for (float datum : data) {
             sum += datum;
         }
-        double mean = sum / (double) data.length;
-        sum = 0.0;
-        double delta;
-        for (double datum : data) {
+        float mean = sum / (float) data.length;
+        sum = 0;
+        float delta;
+        for (float datum : data) {
             delta = datum - mean;
             sum += (delta * delta);
         }
-        double standardDeviation = Math.sqrt(sum / (double) data.length);
+        float standardDeviation = (float) Math.sqrt(sum / (float) data.length);
         for (int i = 0; i < data.length; i++) {
             result[i] = (data[i] - mean) / standardDeviation;
         }
