+++ src/main/java/de/jklotz/neuralnetwork/normalization/MinMax.java
@@ -3,11 +3,11 @@ package de.jklotz.neuralnetwork.normalization;
 public class MinMax implements Normalization {
 
     @Override
-    public double[] normalize(double[] data) {
-        double[] result = new double[data.length];
-        double min = Double.MAX_VALUE;
-        double max = Double.MIN_VALUE;
-        for (double datum : data) {
+    public float[] normalize(float[] data) {
+        float[] result = new float[data.length];
+        float min = Float.MAX_VALUE;
+        float max = Float.MIN_VALUE;
+        for (float datum : data) {
             min = Math.min(min, datum);
             max = Math.max(max, datum);
         }
