+++ src/main/java/de/jklotz/neuralnetwork/normalization/Normalization.java
@@ -7,5 +7,5 @@ public interface Normalization extends Serializable {
     Gaussian GAUSSIAN = new Gaussian();
     MinMax MIN_MAX = new MinMax();
 
-    double[] normalize(double[] data);
+    float[] normalize(float[] data);
 }
